+++ Sources/Models/Async/AsyncRegistration.swift
@@ -16,54 +16,28 @@ struct AsyncRegistration: Sendable {
     let asyncRegistrationFactory: AsyncRegistrationFactory
 
     /// Initializer for registrations that don't need any variable argument
-    init<T: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving) async -> T) {
+    init<Dependency: Sendable>(type: Dependency.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving) async -> Dependency) {
         identifier = RegistrationIdentifier(type: type)
         self.scope = scope
         asyncRegistrationFactory = { resolver, _ in await factory(resolver) }
     }
 
-    /// Initializer for registrations that expect a variable argument passed to the factory closure when the dependency is being resolved
-    init<T: Sendable, Argument: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, Argument) async -> T) {
-        let registrationIdentifier = RegistrationIdentifier(type: type, argument: Argument.self)
+    /// Initializer for registrations that expect variable arguments passed to the factory closure when the dependency is being resolved
+    ///
+    /// Uses Swift parameter packs to support 1-3 arguments with a single initializer.
+    init<Dependency: Sendable, each Argument: Sendable>(type: Dependency.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, repeat each Argument) async -> Dependency) {
+        let registrationIdentifier = RegistrationIdentifier(type: type, argumentTypes: repeat (each Argument).self)
 
         identifier = registrationIdentifier
         self.scope = scope
         asyncRegistrationFactory = { resolver, arg in
-            guard let argument = arg as? Argument else {
-                throw ResolutionError.unmatchingArgumentType(message: "Registration of type \(registrationIdentifier.description) doesn't accept an argument of type \(Argument.self)")
+            guard let arguments = arg as? (repeat each Argument) else {
+                throw ResolutionError.unmatchingArgumentType(
+                    message: "Registration of type \(registrationIdentifier.description) doesn't accept arguments of type \(Swift.type(of: arg))"
+                )
             }
 
-            return await factory(resolver, argument)
-        }
-    }
-
-    /// Initializer for registrations that expect two variable arguments passed to the factory closure when the dependency is being resolved
-    init<T: Sendable, Argument1: Sendable, Argument2: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, Argument1, Argument2) async -> T) {
-        let registrationIdentifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self)
-
-        identifier = registrationIdentifier
-        self.scope = scope
-        asyncRegistrationFactory = { resolver, arg in
-            guard let arguments = arg as? (Argument1, Argument2) else {
-                throw ResolutionError.unmatchingArgumentType(message: "Registration of type \(registrationIdentifier.description) doesn't accept arguments of type (\(Argument1.self), \(Argument2.self))")
-            }
-
-            return await factory(resolver, arguments.0, arguments.1)
-        }
-    }
-
-    /// Initializer for registrations that expect three variable arguments passed to the factory closure when the dependency is being resolved
-    init<T: Sendable, Argument1: Sendable, Argument2: Sendable, Argument3: Sendable>(type: T.Type, scope: DependencyScope, factory: @Sendable @escaping (any AsyncDependencyResolving, Argument1, Argument2, Argument3) async -> T) {
-        let registrationIdentifier = RegistrationIdentifier(type: type, argument1: Argument1.self, argument2: Argument2.self, argument3: Argument3.self)
-
-        identifier = registrationIdentifier
-        self.scope = scope
-        asyncRegistrationFactory = { resolver, arg in
-            guard let arguments = arg as? (Argument1, Argument2, Argument3) else {
-                throw ResolutionError.unmatchingArgumentType(message: "Registration of type \(registrationIdentifier.description) doesn't accept arguments of type (\(Argument1.self), \(Argument2.self), \(Argument3.self))")
-            }
-
-            return await factory(resolver, arguments.0, arguments.1, arguments.2)
+            return await factory(resolver, repeat each arguments)
         }
     }
 }