+++ Sources/Protocols/Registration/Sync/DependencyRegistering.swift
@@ -9,17 +9,11 @@ import Foundation
 
 /// A type that is able to register a dependency
 public protocol DependencyRegistering {
-    /// Factory closure that instantiates the required dependency
+    /// Factory closure that instantiates the required dependency without arguments
     typealias Factory<Dependency> = (DependencyResolving) -> Dependency
 
-    /// Factory closure that instantiates the required dependency with the given variable argument
-    typealias FactoryWithOneArgument<Dependency, Argument> = (DependencyResolving, Argument) -> Dependency
-
-    /// Factory closure that instantiates the required dependency with two given variable arguments
-    typealias FactoryWithTwoArguments<Dependency, Argument1, Argument2> = (DependencyResolving, Argument1, Argument2) -> Dependency
-
-    /// Factory closure that instantiates the required dependency with three given variable arguments
-    typealias FactoryWithThreeArguments<Dependency, Argument1, Argument2, Argument3> = (DependencyResolving, Argument1, Argument2, Argument3) -> Dependency
+    /// Factory closure that instantiates the required dependency with variable arguments (1-3 supported)
+    typealias FactoryWithArguments<Dependency, each Argument> = (DependencyResolving, repeat each Argument) -> Dependency
 
     /// Register a dependency
     ///
@@ -29,19 +23,9 @@ public protocol DependencyRegistering {
     ///   - factory: Closure that is called when the dependency is being resolved
     func register<Dependency>(type: Dependency.Type, in scope: DependencyScope, factory: @escaping Factory<Dependency>)
 
-    /// Register a dependency with a variable argument
+    /// Register a dependency with variable arguments
     ///
-    /// The argument is typically a parameter in an initializer of the dependency that is not registered in the same resolver (i.e. container),
-    /// therefore, it needs to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
-    /// should always return a new instance for dependencies with arguments.
-    ///
-    /// - Parameters:
-    ///   - type: Type of the dependency to register
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument>(type: Dependency.Type, factory: @escaping FactoryWithOneArgument<Dependency, Argument>)
-
-    /// Register a dependency with two variable arguments
-    ///
+    /// Uses Swift parameter packs to support 1-3 arguments with a single method signature.
     /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same resolver (i.e. container),
     /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
     /// should always return a new instance for dependencies with arguments.
@@ -49,38 +33,11 @@ public protocol DependencyRegistering {
     /// - Parameters:
     ///   - type: Type of the dependency to register
     ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument1, Argument2>(type: Dependency.Type, factory: @escaping FactoryWithTwoArguments<Dependency, Argument1, Argument2>)
-
-    /// Register a dependency with three variable arguments
-    ///
-    /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same resolver (i.e. container),
-    /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
-    /// should always return a new instance for dependencies with arguments.
-    ///
-    /// - Parameters:
-    ///   - type: Type of the dependency to register
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument1, Argument2, Argument3>(type: Dependency.Type, factory: @escaping FactoryWithThreeArguments<Dependency, Argument1, Argument2, Argument3>)
+    func register<Dependency, each Argument>(type: Dependency.Type, factory: @escaping FactoryWithArguments<Dependency, repeat each Argument>)
 }
 
 // MARK: Overloaded factory methods
 public extension DependencyRegistering {
-    /// Default ``DependencyScope`` value
-    ///
-    /// The default value is `shared`
-    static var defaultScope: DependencyScope {
-        DependencyScope.shared
-    }
-
-    /// Register a dependency in the default ``DependencyScope``, i.e. in the `shared` scope
-    ///
-    /// - Parameters:
-    ///   - type: Type of the dependency to register
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency>(type: Dependency.Type, factory: @escaping Factory<Dependency>) {
-        register(type: type, in: Self.defaultScope, factory: factory)
-    }
-
     /// Register a dependency with an implicit type determined by the factory closure return type
     ///
     /// - Parameters:
@@ -90,47 +47,16 @@ public extension DependencyRegistering {
         register(type: Dependency.self, in: scope, factory: factory)
     }
 
-    /// Register a dependency with an implicit type determined by the factory closure return type and in the default ``DependencyScope``, i.e. in the `shared` scope
+    /// Register a dependency with variable arguments. The type of the dependency is determined implicitly based on the factory closure return type
     ///
-    /// - Parameters:
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency>(factory: @escaping Factory<Dependency>) {
-        register(type: Dependency.self, in: Self.defaultScope, factory: factory)
-    }
-
-    /// Register a dependency with a variable argument. The type of the dependency is determined implicitly based on the factory closure return type
-    ///
-    /// The argument is typically a parameter in an initializer of the dependency that is not registered in the same resolver (i.e. container),
-    /// therefore, it needs to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
-    /// should always return a new instance for dependencies with arguments.
-    ///
-    /// - Parameters:
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument>(factory: @escaping FactoryWithOneArgument<Dependency, Argument>) {
-        register(type: Dependency.self, factory: factory)
-    }
-
-    /// Register a dependency with two variable arguments. The type of the dependency is determined implicitly based on the factory closure return type
-    ///
+    /// Uses Swift parameter packs to support 1-3 arguments with a single method signature.
     /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same resolver (i.e. container),
     /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
     /// should always return a new instance for dependencies with arguments.
     ///
     /// - Parameters:
     ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument1, Argument2>(factory: @escaping FactoryWithTwoArguments<Dependency, Argument1, Argument2>) {
-        register(type: Dependency.self, factory: factory)
-    }
-
-    /// Register a dependency with three variable arguments. The type of the dependency is determined implicitly based on the factory closure return type
-    ///
-    /// The arguments are typically parameters in an initializer of the dependency that are not registered in the same resolver (i.e. container),
-    /// therefore, they need to be passed in `resolve` call. This registration method doesn't have any scope parameter for a reason - the container
-    /// should always return a new instance for dependencies with arguments.
-    ///
-    /// - Parameters:
-    ///   - factory: Closure that is called when the dependency is being resolved
-    func register<Dependency, Argument1, Argument2, Argument3>(factory: @escaping FactoryWithThreeArguments<Dependency, Argument1, Argument2, Argument3>) {
+    func register<Dependency, each Argument>(factory: @escaping FactoryWithArguments<Dependency, repeat each Argument>) {
         register(type: Dependency.self, factory: factory)
     }
 }